backend/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib==1.7.4
bcrypt==3.2.2
python-multipart